Introduce OnigStringContext for naive string caching.
OnigStringContext is an immutable object that encapsulates
various values which are derived from a v8String, including
a UTF8 copy, a UTF16 copy and a predicate.

OnigStringContext can be used as a cache of these values to
save extra computation like string copying and memory allocation
when the same V8 string is given to FindNextMatchSync() in
subsequent calls.

Speedup:
 * large.js sync:    482ms ->  366ms
 * large.js async:  5914ms -> 4971ms (These numbers are noisy)
 * oneline sync:   66624ms -> 1085ms
 * oneline async: N/A (took too long)

Although there are some rooms for improvement in the async case,
this overall seems good starting point to address slow-long-line
problem, where we've repeatedly allocated the long string.